First draft of the graphical modeling tool.
The user is able to add new tasks to the canvas and drag them around. There is also a context menu, which shows some additional functions. Via the connect-function the user is able to connect two tasks with an arrow. This arrow updates its position automatically, when one of the tasks is moved.