Implement an interface to the server.
Since the workflows are executed completely on the server an interface to transfer the models is needed. When the user selects a previously created model in the tabs, it is requested from the server and saved in the state. All modifications are done on the internal state. Only when the user clicks the save button the updated model in the state is transferred to the server.