Update to the RAML form.
Instead of entering the filename and the respective resource manually, the user is able to select the file or resource from a set of parsed raml files. To be able to do so, the files have to be uploaded first. During this process, the raml files are processed on the server and a list of resources is returned.
These "api-descriptions" are also managed in the state by the apiSlice.
Also changed the overall data flow (e.g. by limiting the model state to a single active model).