Implement the switch form & a new form evaluation approach.
Instead of evaluating a giant form with all possible parameters, I create a separate subform for each task type (invoke, assign, switch). These forms are evaluated when the user clicks on the "update model" button. During this process the different forms are merged together and the resulting task is saved to the state.